% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webvitals.R
\name{implement_webvitals}
\alias{implement_webvitals}
\title{Console log and query the CLS, FID and LCP using the web-vitals JS library}
\usage{
implement_webvitals(FID_id, LCP_id, CLS_id, wait_to_get_info = 60000)
}
\arguments{
\item{FID_id}{the ID of the FID value that will be collected}

\item{LCP_id}{the ID of the LCP value that will be collected}

\item{CLS_id}{the ID of the CLS value that will be collected}

\item{wait_to_get_info}{how long (in milliseconds) do you want to wait before collecting the web-vitals metrics. Defaults to 6000ms}
}
\value{
called for the side effect of using the web-vitals JS library.
}
\description{
Console log and query the CLS, FID and LCP using the web-vitals JS library
}
